+++ env.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-undef */
 /* eslint-disable no-useless-escape */
+
 const os = require('os');
 
 const playwright = require('playwright-core');
@@ -10,7 +11,8 @@ const pptrVersion = pptrPackageJSON.version;
 const packageJson = require('./package.json');
 const IS_DOCKER = process.env.IS_DOCKER === 'true';
 
-const USE_CHROME_STABLE = process.env.USE_CHROME_STABLE && process.env.USE_CHROME_STABLE === 'true';
+const USE_CHROME_STABLE =
+  process.env.USE_CHROME_STABLE && process.env.USE_CHROME_STABLE === 'true';
 
 const MAC = 'MAC';
 const WINDOWS = 'WINDOWS';
@@ -23,13 +25,14 @@ const CHROME_BINARY_PATHS = {
   WIN: 'C:\Program Files (x86)\Google\Chrome\Application\chrome.exe',
 };
 
-const PLATFORM = os.platform() === 'win32' ?
-  WINDOWS :
-    os.platform() === 'darwin' ?
-      MAC :
-      os.arch() === 'arm64' ?
-        LINUX_ARM64 :
-        LINUX;
+const PLATFORM =
+  os.platform() === 'win32'
+    ? WINDOWS
+    : os.platform() === 'darwin'
+    ? MAC
+    : os.arch() === 'arm64'
+    ? LINUX_ARM64
+    : LINUX;
 
 /*
  * Assess which chromium revision to install.
@@ -55,12 +58,12 @@ const PUPPETEER_CHROMIUM_REVISION = (() => {
     return pptrPackageJSON.puppeteer.chromium_revision;
   }
 
-
   if (puppeteer._preferredRevision) {
     return puppeteer._preferredRevision;
   }
 
-  return require('puppeteer/lib/cjs/revisions').PUPPETEER_REVISIONS.chromium;
+  const pptr = require('./node_modules/puppeteer/node_modules/puppeteer-core/lib/cjs/puppeteer/revisions');
+  return pptr.PUPPETEER_REVISIONS.chromium;
 })();
 
 /*
@@ -72,8 +75,10 @@ const PUPPETEER_BINARY_LOCATION = (() => {
     return playwright.chromium.executablePath();
   }
 
-  const browserFetcher = puppeteer.createBrowserFetcher();
-  return browserFetcher.revisionInfo(PUPPETEER_CHROMIUM_REVISION).executablePath;
+  const browserFetcher = puppeteer.createBrowserFetcher({ product: 'chrome' });
+
+  return browserFetcher.revisionInfo(PUPPETEER_CHROMIUM_REVISION)
+    .executablePath;
 })();
 
 /*