+++ scripts/postinstall.js
@@ -9,9 +9,7 @@ const extract = require('extract-zip');
 const fs = require('fs-extra');
 const _ = require('lodash');
 const fetch = require('node-fetch');
-const {
-  installBrowsersForNpmInstall,
-} = require('playwright-core/lib/server');
+const { installBrowsersForNpmInstall } = require('playwright-core/lib/server');
 const puppeteer = require('puppeteer');
 const rimraf = require('rimraf');
 
@@ -134,7 +132,9 @@ const downloadChromium = () => {
     `Downloading chromium for revision ${PUPPETEER_CHROMIUM_REVISION}`,
   );
 
-  return puppeteer.createBrowserFetcher().download(PUPPETEER_CHROMIUM_REVISION);
+  return puppeteer
+    .createBrowserFetcher({ product: 'chrome' })
+    .download(PUPPETEER_CHROMIUM_REVISION);
 };
 
 const downloadChromedriver = () => {
@@ -239,10 +239,11 @@ const downloadDevTools = () => {
     } finally {
       rimraf(browserlessTmpDir, (err) => {
         console.log('Done unpacking chromedriver and devtools assets');
-        if (err)
+        if (err) {
           console.warn(
             `Error removing temporary directory ${browserlessTmpDir}`,
           );
+        }
         resolve();
       });
     }