+++ src/tests/integrations/webdriver.spec.ts
@@ -1,3 +1,5 @@
+import { expect } from 'chai';
+
 import { BrowserlessServer } from '../../browserless';
 import { IBrowserlessOptions } from '../../types.d';
 import { sleep } from '../../utils';
@@ -17,6 +19,7 @@ describe('Browserless Chrome Webdriver', () => {
     (browserless = new BrowserlessServer(args));
 
   afterEach(async () => {
+    sleep(1000);
     await browserless.kill();
   });
 
@@ -45,9 +48,9 @@ describe('Browserless Chrome Webdriver', () => {
     await Promise.all([run(), run()]);
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(2);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(2);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('handles driver close calls', async () => {
@@ -72,9 +75,9 @@ describe('Browserless Chrome Webdriver', () => {
     await run();
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(1);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('runs lengthy sessions', async () => {
@@ -108,9 +111,9 @@ describe('Browserless Chrome Webdriver', () => {
     await Promise.all([run()]);
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(1);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('works with no timeouts', async () => {
@@ -138,10 +141,10 @@ describe('Browserless Chrome Webdriver', () => {
     await run();
     await sleep(50);
 
-    expect(browserless.currentStat.timedout).toEqual(0);
-    expect(browserless.currentStat.successful).toEqual(1);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.timedout).to.equal(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('works with job-based timeouts', async () => {
@@ -154,7 +157,7 @@ describe('Browserless Chrome Webdriver', () => {
 
     await browserless.startServer();
     chromeCapabilities.set('goog:chromeOptions', webdriverOpts);
-    chromeCapabilities.set('browserless.timeout', 10);
+    chromeCapabilities.set('browserless:timeout', 10);
 
     async function run() {
       const driver = new webdriver.Builder()
@@ -168,12 +171,12 @@ describe('Browserless Chrome Webdriver', () => {
     }
 
     await run();
-    await sleep(50);
+    await sleep(1000);
 
-    expect(browserless.currentStat.timedout).toEqual(1);
-    expect(browserless.currentStat.successful).toEqual(0);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.timedout).to.equal(1);
+    expect(browserless.currentStat.successful).to.equal(0);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('authorizes with tokens', async () => {
@@ -201,9 +204,9 @@ describe('Browserless Chrome Webdriver', () => {
     await run();
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(1);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('authorizes with webdriver-backed tokens', async () => {
@@ -229,22 +232,15 @@ describe('Browserless Chrome Webdriver', () => {
     {
       const chromeCapabilities = webdriver.Capabilities.chrome();
       chromeCapabilities.set('goog:chromeOptions', webdriverOpts);
-      chromeCapabilities.set('browserless.token', 'abcd');
-      await run(chromeCapabilities);
-    }
-
-    {
-      const chromeCapabilities = webdriver.Capabilities.chrome();
-      chromeCapabilities.set('goog:chromeOptions', webdriverOpts);
       chromeCapabilities.set('browserless:token', 'abcd');
       await run(chromeCapabilities);
     }
 
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(2);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
   });
 
   it('queues sessions', async () => {
@@ -269,8 +265,8 @@ describe('Browserless Chrome Webdriver', () => {
     await Promise.all([run(), run()]);
     await sleep();
 
-    expect(browserless.currentStat.successful).toEqual(2);
-    expect(browserless.currentStat.queued).toEqual(1);
+    expect(browserless.currentStat.successful).to.equal(2);
+    expect(browserless.currentStat.queued).to.equal(1);
   });
 
   it('fails requests', async () => {
@@ -292,7 +288,9 @@ describe('Browserless Chrome Webdriver', () => {
       .build()
       .then(throws)
       .catch((error: Error) => {
-        expect(error.message).toEqual('Unable to parse new session response: ');
+        expect(error.message).to.equal(
+          'Unable to parse new session response: ',
+        );
       });
   });
 
@@ -314,7 +312,7 @@ describe('Browserless Chrome Webdriver', () => {
       .build()
       .then(throws)
       .catch((error: Error) => {
-        expect(error.message).toEqual('Unauthorized');
+        expect(error.message).to.equal('Unauthorized');
       });
   });
 
@@ -340,13 +338,13 @@ describe('Browserless Chrome Webdriver', () => {
 
     await sleep(50);
 
-    expect(browserless.currentStat.successful).toEqual(1);
-    expect(browserless.currentStat.rejected).toEqual(0);
-    expect(browserless.currentStat.queued).toEqual(0);
+    expect(browserless.currentStat.successful).to.equal(1);
+    expect(browserless.currentStat.rejected).to.equal(0);
+    expect(browserless.currentStat.queued).to.equal(0);
     const processes = await getChromeProcesses();
 
     await sleep(50);
 
-    expect(processes.stdout).not.toContain('.local-chromium');
+    expect(processes.stdout).to.not.contain('.local-chromium');
   });
 });