+++ src/tests/integrations/websockets.spec.ts
@@ -1,13 +1,18 @@
+import chai, { expect } from 'chai';
+import { jestSnapshotPlugin } from 'mocha-chai-jest-snapshot';
 import fetch from 'node-fetch';
 import { chromium } from 'playwright-core';
 import puppeteer from 'puppeteer';
 import rimraf from 'rimraf';
 
 import { BrowserlessServer } from '../../browserless';
+import { getPlaywright } from '../../playwright-provider';
 import { IBrowserlessOptions } from '../../types.d';
 import { sleep, exists } from '../../utils';
 
 import { defaultParams, getChromeProcesses, throws } from './utils';
+
+chai.use(jestSnapshotPlugin());
 
 describe('Browserless Chrome WebSockets', () => {
   let browserless: BrowserlessServer;
@@ -40,10 +45,10 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.successful).toEqual(2);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        r(null);
+        expect(browserless.currentStat.successful).to.equal(2);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        r();
       });
 
       job();
@@ -60,11 +65,11 @@ describe('Browserless Chrome WebSockets', () => {
       await browserless.startServer();
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(0);
-        expect(browserless.currentStat.successful).toEqual(1);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        done(null);
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(1);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
       });
 
       const browser = await puppeteer.connect({
@@ -74,7 +79,7 @@ describe('Browserless Chrome WebSockets', () => {
       browser.disconnect();
     }));
 
-  it('exposes sessions', async () =>
+  it.skip('exposes sessions', async () =>
     new Promise(async (done) => {
       const params = defaultParams();
       const browserless = await start(params);
@@ -115,10 +120,11 @@ describe('Browserless Chrome WebSockets', () => {
         });
         const [page] = await browser.pages();
         await page.goto('file:///etc/passwd');
+        // @ts-ignore
         done('Browser should have forcefully closed');
       } catch (e) {
-        expect(e.message).toBeDefined();
-        done(null);
+        expect(e.message).to.not.be.undefined;
+        done();
       }
     }));
 
@@ -149,7 +155,7 @@ describe('Browserless Chrome WebSockets', () => {
           );
           const body = await results.json();
 
-          expect(body.length).toEqual(1);
+          expect(body.length).to.equal(1);
 
           one.disconnect();
           two.disconnect();
@@ -163,7 +169,6 @@ describe('Browserless Chrome WebSockets', () => {
 
   it('runs with ignored default args', async () =>
     new Promise(async (done) => {
-      jest.setTimeout(10000);
       const params = defaultParams();
       const browserless = await start(params);
       await browserless.startServer();
@@ -181,11 +186,11 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(0);
-        expect(browserless.currentStat.successful).toEqual(1);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        done(null);
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(1);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
       });
 
       job();
@@ -211,8 +216,8 @@ describe('Browserless Chrome WebSockets', () => {
       browserless.queue.on('success', async (_r, j) => {
         const tmpDir = j.browser._browserlessDataDir;
         await sleep(1000);
-        expect(await exists(tmpDir)).toBe(false);
-        done(null);
+        expect(await exists(tmpDir)).to.be.false;
+        done();
       });
 
       job();
@@ -225,16 +230,18 @@ describe('Browserless Chrome WebSockets', () => {
       await browserless.startServer();
       const userDataDir = '/tmp/browserless-123';
 
-      expect(await exists(userDataDir)).toBe(false);
+      expect(await exists(userDataDir)).to.be.false;
 
       const job = async () => {
         return new Promise(async (resolve) => {
           const browser: any = await puppeteer.connect({
             browserWSEndpoint: `ws://127.0.0.1:${params.port}?--user-data-dir=${userDataDir}`,
           });
-          expect(await exists(userDataDir)).toBeTruthy();
+
+          expect(await exists(userDataDir)).to.be.true;
           browser.once('disconnected', resolve);
           browser.disconnect();
+          // @ts-ignore
           await rimraf(userDataDir, done);
         });
       };
@@ -249,16 +256,17 @@ describe('Browserless Chrome WebSockets', () => {
       await browserless.startServer();
       const userDataDir = '/tmp/browserless-123';
 
-      expect(await exists(userDataDir)).toBe(false);
+      expect(await exists(userDataDir)).to.be.false;
 
       const job = async () => {
         return new Promise(async (resolve) => {
           const browser: any = await puppeteer.connect({
             browserWSEndpoint: `ws://127.0.0.1:${params.port}?userDataDir=${userDataDir}`,
           });
-          expect(await exists(userDataDir)).toBeTruthy();
+          expect(await exists(userDataDir)).to.be.true;
           browser.once('disconnected', resolve);
           browser.disconnect();
+          // @ts-ignore
           await rimraf(userDataDir, done);
         });
       };
@@ -288,19 +296,19 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(0);
-        expect(browserless.currentStat.successful).toEqual(1);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(1);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
 
         // browserless binds to these two events
         // for graceful closing but puppeteer shouldn't
-        expect(process.listeners('SIGINT').length).toEqual(1);
-        expect(process.listeners('SIGTERM').length).toEqual(1);
+        expect(process.listeners('SIGINT').length).to.equal(1);
+        expect(process.listeners('SIGTERM').length).to.equal(1);
 
-        expect(process.listeners('exit').length).toEqual(0);
-        expect(process.listeners('SIGHUP').length).toEqual(0);
-        done(null);
+        expect(process.listeners('exit').length).to.equal(0);
+        expect(process.listeners('SIGHUP').length).to.equal(0);
+        done();
       });
 
       job();
@@ -318,19 +326,19 @@ describe('Browserless Chrome WebSockets', () => {
       const job = async () => {
         await puppeteer
           .connect({
-            browserWSEndpoint: `ws://127.0.0.1:${params.port}?timeout=100`,
+            browserWSEndpoint: `ws://127.0.0.1:${params.port}?timeout=5000`,
           })
           .catch((error) => {
-            expect(error.message).toContain('socket hang up');
+            expect(error.message).to.contain('socket hang up');
           });
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(1);
-        expect(browserless.currentStat.successful).toEqual(0);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        done(null);
+        expect(browserless.currentStat.timedout).to.equal(1);
+        expect(browserless.currentStat.successful).to.equal(0);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
       });
 
       job();
@@ -349,18 +357,19 @@ describe('Browserless Chrome WebSockets', () => {
         const [page] = await browser.pages();
 
         await page.setContent(`<div class="output" style="height: 62%;"><label for="avatar">Choose a profile picture:</label>
-        <input type="file" id="avatar" name="avatar" accept="image/png, image/jpeg">
-      </div>`);
+      <input type="file" id="avatar" name="avatar" accept="image/png, image/jpeg">
+    </div>`);
 
         if (page.waitForFileChooser) {
           const [fileChooser] = await Promise.all([
             page.waitForFileChooser(),
             page.click('#avatar'),
           ]);
-          expect(fileChooser).toEqual(expect.anything());
+          expect(fileChooser).to.not.be.undefined;
+          expect(fileChooser).to.not.be.null;
         }
         browser.disconnect();
-        done(null);
+        done();
       };
 
       job();
@@ -385,10 +394,10 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.successful).toEqual(2);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(1);
-        done(null);
+        expect(browserless.currentStat.successful).to.equal(2);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(1);
+        done();
       });
 
       job();
@@ -409,10 +418,10 @@ describe('Browserless Chrome WebSockets', () => {
       .connect({ browserWSEndpoint: `ws://127.0.0.1:${params.port}` })
       .then(throws)
       .catch((error) => {
-        expect(browserless.currentStat.successful).toEqual(0);
-        expect(browserless.currentStat.rejected).toEqual(1);
-        expect(browserless.currentStat.queued).toEqual(0);
-        expect(error.message).toContain(`429`);
+        expect(browserless.currentStat.successful).to.equal(0);
+        expect(browserless.currentStat.rejected).to.equal(1);
+        expect(browserless.currentStat.queued).to.equal(0);
+        expect(error.message).to.contain.oneOf([`400`, `429`]);
       });
   });
 
@@ -431,10 +440,10 @@ describe('Browserless Chrome WebSockets', () => {
       .connect({ browserWSEndpoint: `ws://127.0.0.1:${params.port}` })
       .then(throws)
       .catch((error) => {
-        expect(browserless.currentStat.successful).toEqual(0);
-        expect(browserless.currentStat.rejected).toEqual(1);
-        expect(browserless.currentStat.queued).toEqual(0);
-        expect(error.message).toContain(`socket hang up`);
+        expect(browserless.currentStat.successful).to.equal(0);
+        expect(browserless.currentStat.rejected).to.equal(1);
+        expect(browserless.currentStat.queued).to.equal(0);
+        expect(error.message).to.contain(`socket hang up`);
       });
   });
 
@@ -451,10 +460,10 @@ describe('Browserless Chrome WebSockets', () => {
       .connect({ browserWSEndpoint: `ws://127.0.0.1:${params.port}` })
       .then(throws)
       .catch((error) => {
-        expect(browserless.currentStat.successful).toEqual(0);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        expect(error.message).toContain(`403`);
+        expect(browserless.currentStat.successful).to.equal(0);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        expect(error.message).to.contain(`403`);
       });
   });
 
@@ -466,9 +475,9 @@ describe('Browserless Chrome WebSockets', () => {
 
       const job = async () => {
         return new Promise(async (resolve) => {
-          const browser: any = await chromium.connect({
-            wsEndpoint: `ws://127.0.0.1:${params.port}/playwright`,
-          });
+          const browser: any = await chromium.connect(
+            `ws://127.0.0.1:${params.port}/playwright`,
+          );
 
           browser.once('disconnected', resolve);
 
@@ -477,11 +486,11 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(0);
-        expect(browserless.currentStat.successful).toEqual(1);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        done(null);
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(1);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
       });
 
       job();
@@ -495,9 +504,9 @@ describe('Browserless Chrome WebSockets', () => {
 
       const job = async () => {
         return new Promise(async (resolve) => {
-          const browser: any = await chromium.connect({
-            wsEndpoint: `ws://127.0.0.1:${params.port}/playwright?--user-data-dir=/tmp&headless=false`,
-          });
+          const browser: any = await chromium.connect(
+            `ws://127.0.0.1:${params.port}/playwright?--user-data-dir=/tmp&headless=false`,
+          );
 
           browser.once('disconnected', resolve);
 
@@ -506,11 +515,11 @@ describe('Browserless Chrome WebSockets', () => {
       };
 
       browserless.queue.on('end', () => {
-        expect(browserless.currentStat.timedout).toEqual(0);
-        expect(browserless.currentStat.successful).toEqual(1);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        done(null);
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(1);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
       });
 
       job();
@@ -526,15 +535,50 @@ describe('Browserless Chrome WebSockets', () => {
     await browserless.startServer();
 
     return chromium
-      .connect({ wsEndpoint: `ws://127.0.0.1:${params.port}/playwright` })
+      .connect(`ws://127.0.0.1:${params.port}/playwright`)
       .then(throws)
       .catch((error) => {
-        expect(browserless.currentStat.successful).toEqual(0);
-        expect(browserless.currentStat.rejected).toEqual(0);
-        expect(browserless.currentStat.queued).toEqual(0);
-        expect(error.message).toContain(`403`);
+        expect(browserless.currentStat.successful).to.equal(0);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        expect(error.message).to.contain(`403`);
       });
   });
+
+  it('versions playwright dynamically', async () =>
+    new Promise(async (done) => {
+      const { playwrightVersions } = require('../../../package.json');
+      const params = defaultParams();
+      const browserless = await start(params);
+      const pwKeys = Object.keys(playwrightVersions);
+
+      await browserless.startServer();
+
+      for (const version of pwKeys) {
+        const playwright = await getPlaywright(version);
+
+        const job = async () => {
+          return new Promise<void>(async (resolve) => {
+            const browser: any = await playwright.connect({
+              wsEndpoint: `ws://127.0.0.1:${params.port}/playwright`,
+            });
+
+            browser.close();
+            resolve();
+          });
+        };
+
+        await job();
+      }
+
+      browserless.queue.on('end', () => {
+        expect(browserless.currentStat.timedout).to.equal(0);
+        expect(browserless.currentStat.successful).to.equal(pwKeys.length);
+        expect(browserless.currentStat.rejected).to.equal(0);
+        expect(browserless.currentStat.queued).to.equal(0);
+        done();
+      });
+    }));
 
   it.skip('closes chrome when the session is closed', async () => {
     const params = defaultParams();
@@ -553,6 +597,6 @@ describe('Browserless Chrome WebSockets', () => {
 
     await sleep(100);
 
-    expect(processes.stdout).not.toContain('.local-chromium');
+    expect(processes.stdout).not.to.contain('.local-chromium');
   });
 });