+++ CHANGELOG.md
@@ -1,5 +1,53 @@
-# [Latest](https://github.com/browserless/chrome/compare/v1.53.0...master)
+# [Latest](https://github.com/browserless/chrome/compare/v1.57.0...master)
 - Dependency updates.
+
+# [v1.57.0](https://github.com/browserless/chrome/compare/v1.56.0...v1.57.0)
+- Dependency updates.
+- Add back in `puppeteer@13.1.3` for PDF/Screenshot performance.
+- Replace `puppeteer@18.0.5` in favor of `puppeteer@19.2.2`.
+- Better puppeteer install/binary path lookups.
+- Support for the following playwright versions via the /playwright path:
+  - 1.23.3
+  - 1.24.2
+  - 1.25.2
+  - 1.26.1
+  - 1.27.1
+
+# [v1.56.0](https://github.com/browserless/chrome/compare/v1.55.0...v1.56.0)
+- Dependency updates.
+- Fixes various puppeteer issues with regards to revision information and browser-fetching.
+- Fixes a Selenium issue where failed browsers can cause session to zombie.
+- Small fix for route error handling.
+- Better selenium W3C support.
+
+# [v1.55.0](https://github.com/browserless/chrome/compare/v1.55.0...v1.54.1)
+**Potentially Breaking**
+- Update base OS from Ubuntu 20.04 to Ubuntu 22.04.
+- Drops `path` support for both `addScriptTag` and `addStyleTag`. Use `url` or inject content directly.
+---
+- Dependency updates.
+- We now support multiple versions of playwright via `playwright-core`. browserless checks the User-Agent string and tries to load a compatible playwright server at runtime in order to avoid breakages between versions.
+- New `scrollPage` boolean parameter for the screenshot and other relevant APIs.
+- browserless now writes appropriately formatted HTTP and WebSocket messages for non-2xx responses.
+- Better CI coverage with prior puppeteer versions.
+- Adds back a `browser.json` file in our docker images for easier CDP-related queries.
+- A new `PORT` arg is available at build time in order to specify a PORT to listen onto.
+- Updated links for browserless documentation in docs.
+
+# [v1.54.1](https://github.com/browserless/chrome/compare/v1.54.1...v1.54.0)
+- Dependency updates.
+- Add `browser.json` file to git tags.
+- Prettier fixes on source.
+- deploy.js file fixes for creating production tags.
+- Fix: add the `userDataDir` option so that we don't fill disks when using `stealth`.
+- Fix: allow `disconnect` listeners to fire before closing.
+
+# [v1.54.0](https://github.com/browserless/chrome/compare/v1.53.0...v1.54.0)
+- Dependency updates.
+- Move to using `mocha` for better esm support.
+- Drop puppeteer `13.6.0` in favor of `14.4.1`.
+- Utilize puppeteer `14.4.1` for `chrome-stable`.
+- New internal `getCDPClient` for loading a page's CPD connection.
 
 # [v1.53.0](https://github.com/browserless/chrome/compare/v1.52.1...v1.53.0)
 - Dependency updates.